+++ cmd/forklift/stage/bundles.go
@@ -2,6 +2,8 @@ package stage
 
 import (
 	"fmt"
+	"os"
+	"path/filepath"
 	"slices"
 	"strings"
 
@@ -12,7 +14,7 @@ import (
 	fcli "github.com/PlanktoScope/forklift/internal/app/forklift/cli"
 )
 
-// ls-bundle
+// ls-bun
 
 func lsBunAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
@@ -101,5 +103,69 @@ func showBunAction(versions Versions) cli.ActionFunc {
 		}
 		fcli.PrintStagedBundle(0, store, bundle, index, getBundleNames(store)[index])
 		return nil
+	}
+}
+
+// rm-bun
+
+func rmBunAction(versions Versions) cli.ActionFunc {
+	return func(c *cli.Context) error {
+		store, err := getStageStore(c.String("workspace"), versions)
+		if err != nil {
+			return err
+		}
+		if !store.Exists() {
+			return errMissingStore
+		}
+
+		deleteIndex, err := resolveBundleIdentifier(c.Args().First(), store)
+		if err != nil {
+			return err
+		}
+		fmt.Printf("Staged pallet bundle %d will be deleted...\n", deleteIndex)
+		preNext, preHasNext := store.GetNext()
+		preCurrent, preHasCurrent := store.GetCurrent()
+
+		store.RemoveBundleNames(deleteIndex)
+		store.RemoveBundleHistory(deleteIndex)
+		postCurrent, postHasCurrent := store.GetCurrent()
+		switch {
+		case preHasCurrent && !postHasCurrent:
+			fmt.Println(
+				"Warning: now there will be no staged pallet bundle known to have been successfully " +
+					"applied!",
+			)
+		case postHasCurrent && postCurrent != preCurrent:
+			fmt.Printf(
+				"The last staged pallet bundle known to have been successfully applied will change "+
+					"from %d to %d!\n",
+				preCurrent, postCurrent,
+			)
+		}
+		switch {
+		case preHasNext && (preNext == deleteIndex) && postHasCurrent:
+			store.SetNext(postCurrent)
+			fmt.Printf(
+				"Because pallet bundle %d will be deleted, %d will now be the next staged pallet bundle "+
+					"to be applied!\n",
+				deleteIndex, postCurrent,
+			)
+		case preHasNext && (preNext == deleteIndex) && !postHasCurrent:
+			store.SetNext(0)
+			fmt.Printf(
+				"Because bundle %d will be deleted, and there is no remaining successfully-applied pallet "+
+					"bundle in the store's history, now no bundle will be applied next!\n",
+				deleteIndex,
+			)
+		}
+		fmt.Println("Saving the updated state of the stage store...")
+		// Note: we commit the state before deleting the stage (rather than the other way around)
+		// because it's better to accidentally leave the stage on the filesystem than to have indices
+		// of deleted bundles in our history/names/next-state.
+		if err = store.CommitState(); err != nil {
+			return errors.Wrap(err, "couldn't commit the stage store's new state!")
+		}
+		fmt.Println("Deleting the staged pallet bundle from the filesystem...")
+		return os.RemoveAll(filepath.FromSlash(store.GetBundlePath(deleteIndex)))
 	}
 }